% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynhung.r
\name{optimal_cost}
\alias{optimal_cost}
\title{Optimal Cost}
\usage{
optimal_cost(AP)
}
\arguments{
\item{AP}{an object of the assignment_problem class produced with solve_assignment()}
}
\description{
Solve the linear sum assignment problem using the Hungarian method.
}
\examples{
optimal_cost()
}
\keyword{assignment}
\keyword{problem}
