% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynhung.r
\name{optimal_matching}
\alias{optimal_matching}
\title{Optimal Matching}
\usage{
optimal_matching(AP)
}
\arguments{
\item{AP}{an object of the assignment_problem class produced with solve_assignment()}
}
\description{
Solve the linear sum assignment problem using the Hungarian method.
}
\examples{
optimal_matching()
}
\keyword{assignment}
\keyword{problem}
