% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynhung.r
\name{update_cols}
\alias{update_cols}
\title{Update columns of the cost matrix for the assignment problem}
\usage{
update_cols(AP, cost_matrix, cols_changed)
}
\arguments{
\item{AP}{an object of the assignment_problem class produced with solve_assignment()}

\item{cost_matrix}{a new cost matrix}

\item{cols_changed}{a vector of column indices}
}
\description{
Solve the linear sum assignment problem using the Hungarian method.
}
\examples{
update_cols()
}
\keyword{assignment}
\keyword{problem}
