% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynhung.r
\name{update_rows}
\alias{update_rows}
\title{Update rows of the cost matrix for the assignment problem}
\usage{
update_rows(AP, cost_matrix, rows_changed)
}
\arguments{
\item{AP}{an object of the assignment_problem class produced with solve_assignment()}

\item{cost_matrix}{a new cost matrix}

\item{rows_changed}{a vector of row indices}
}
\description{
Solve the linear sum assignment problem using the Hungarian method.
}
\examples{
update_rows()
}
\keyword{assignment}
\keyword{problem}
